Cleanup warnings (#294)
* Silence OpenGL macOS warnings

* Remove version variable from all read and write functions

* Add default case handling to all switch statements

* Remove 'const' type qualifiers that have no effect

* Remove communicator variable name where not required

* Remove unused functions

* Remove all unused parameters

* Remove unused variables

* Properly order initialisation lists

* Fix multi-line comment

* Clear ambiguity of logical operations

* Add missing virtual destructors

* Replace old finite() with isfinite()

* Remove the register specifier

* Change __max to std::max

* Use std::move for SalaErrors in SalaProgram

* Fix operator order

* Fix other minor issues

* Squash all warnings when comparing integers of different signs and sizes

* Read count as an integer

* Remove unused variable

* Remove last version signifiers

* Reimplement various glu functions

Apple's deprecation of OpenGL creates a different problem here because the 3DView is coded using the old OpenGL way. Once 3DView is adapted to work as the GLView (or merged into it) then the glureimpl.h file won't be required. The license of the original file and thus the code in glureimpl.h is the SGI Free Sofware License B, version 2 and it is compatible with the GPL according to https://www.gnu.org/licenses/license-list.html

* Remove unnecessary code

* Remove unused functions and reduntant file

* Add missing math declarations

* Remove non-existant file reference

* Squash some gcc warnings

* Restore and expose a hidden colourscale

The colourscale g_nicecolorhsb along with the relevant comment "Test a range designed to try to keep consistent saturation and brightness of g_nicecolor, and only move hue" has been named "Hue Only Axmanesque" as it's a variation of that colourscale. On the interface it appears as "Equal Ranges (3-Colour Hue Only)"

* Remove qt5_wrap_cpp as AUTOMOC is set to ON

+++ salalib/importutils.cpp
@@ -346,22 +346,22 @@ namespace depthmapX {
         std::vector<Line> lines;
         std::vector<Polyline> polylines;
 
-        for (int jp = 0; jp < dxfLayer.numPoints(); jp++) {
+        for (size_t jp = 0; jp < dxfLayer.numPoints(); jp++) {
             const DxfVertex &dxf_point = dxfLayer.getPoint(jp);
             points.push_back(Point2f(dxf_point.x, dxf_point.y));
         }
 
-        for (int j = 0; j < dxfLayer.numLines(); j++) {
+        for (size_t j = 0; j < dxfLayer.numLines(); j++) {
             const DxfLine &dxf_line = dxfLayer.getLine(j);
             Line line = Line(Point2f(dxf_line.getStart().x, dxf_line.getStart().y),
                              Point2f(dxf_line.getEnd().x, dxf_line.getEnd().y));
             lines.push_back(line);
         }
 
-        for (int k = 0; k < dxfLayer.numPolyLines(); k++) {
+        for (size_t k = 0; k < dxfLayer.numPolyLines(); k++) {
             const DxfPolyLine &poly = dxfLayer.getPolyLine(k);
             std::vector<Point2f> vertices;
-            for (int m = 0; m < poly.numVertices(); m++) {
+            for (size_t m = 0; m < poly.numVertices(); m++) {
                 DxfVertex v = poly.getVertex(m);
                 vertices.push_back(Point2f(v.x, v.y));
             }
@@ -369,10 +369,10 @@ namespace depthmapX {
                 depthmapX::Polyline(vertices, (poly.getAttributes() & DxfPolyLine::CLOSED) == DxfPolyLine::CLOSED));
         }
 
-        for (int l = 0; l < dxfLayer.numSplines(); l++) {
+        for (size_t l = 0; l < dxfLayer.numSplines(); l++) {
             const DxfSpline &poly = dxfLayer.getSpline(l);
             std::vector<Point2f> vertices;
-            for (int m = 0; m < poly.numVertices(); m++) {
+            for (size_t m = 0; m < poly.numVertices(); m++) {
                 DxfVertex v = poly.getVertex(m);
                 vertices.push_back(Point2f(v.x, v.y));
             }
@@ -380,12 +380,12 @@ namespace depthmapX {
                 depthmapX::Polyline(vertices, (poly.getAttributes() & DxfPolyLine::CLOSED) == DxfPolyLine::CLOSED));
         }
 
-        for (int n = 0; n < dxfLayer.numArcs(); n++) {
+        for (size_t n = 0; n < dxfLayer.numArcs(); n++) {
             const DxfArc &circ = dxfLayer.getArc(n);
             std::vector<Point2f> vertices;
-            int segments = circ.numSegments(DXFCIRCLERES);
+            size_t segments = circ.numSegments(DXFCIRCLERES);
             if (segments > 1) {
-                for (int m = 0; m <= segments; m++) {
+                for (size_t m = 0; m <= segments; m++) {
                     DxfVertex v = circ.getVertex(m, segments);
                     vertices.push_back(Point2f(v.x, v.y));
                 }
@@ -393,12 +393,12 @@ namespace depthmapX {
             polylines.push_back(depthmapX::Polyline(vertices, false));
         }
 
-        for (int n = 0; n < dxfLayer.numEllipses(); n++) {
+        for (size_t n = 0; n < dxfLayer.numEllipses(); n++) {
             const DxfEllipse &ellipse = dxfLayer.getEllipse(n);
             std::vector<Point2f> vertices;
-            int segments = ellipse.numSegments(DXFCIRCLERES);
+            size_t segments = ellipse.numSegments(DXFCIRCLERES);
             if (segments > 1) {
-                for (int m = 0; m <= segments; m++) {
+                for (size_t m = 0; m <= segments; m++) {
                     DxfVertex v = ellipse.getVertex(m, segments);
                     vertices.push_back(Point2f(v.x, v.y));
                 }
@@ -406,7 +406,7 @@ namespace depthmapX {
             polylines.push_back(depthmapX::Polyline(vertices, false));
         }
 
-        for (int nc = 0; nc < dxfLayer.numCircles(); nc++) {
+        for (size_t nc = 0; nc < dxfLayer.numCircles(); nc++) {
             const DxfCircle &circ = dxfLayer.getCircle(nc);
             std::vector<Point2f> vertices;
             for (int m = 0; m < DXFCIRCLERES; m++) {
